Automatically infer partitions for hive-style partitioned files
