+++ crates/runtime/src/dataconnector/ftp.rs
@@ -22,9 +22,10 @@ use std::sync::Arc;
 use std::{any::Any, collections::HashMap};
 use url::Url;
 
+use super::listing;
 use super::{
-    DataConnector, DataConnectorFactory, DataConnectorResult, ListingTableConnector, ParameterSpec,
-    Parameters,
+    listing::ListingTableConnector, DataConnector, DataConnectorFactory, DataConnectorResult,
+    ParameterSpec, Parameters,
 };
 
 pub struct FTP {
@@ -113,7 +114,7 @@ impl ListingTableConnector for FTP {
                     message: format!("{} is not a valid URL", dataset.from),
                 })?;
 
-        ftp_url.set_fragment(Some(&super::build_fragments(
+        ftp_url.set_fragment(Some(&listing::build_fragments(
             &self.params,
             vec!["port", "user", "pass", "client_timeout"],
         )));
